requirements.txt
Jinja2==2.10
MarkupSafe==1.0
python-frontmatter==0.4.2
PyYAML==3.12
six==1.11.0
